+++ src/infrastructure/flows/mod.rs
@@ -5,12 +5,13 @@ mod archive_ratings;
 mod change_user_role;
 mod create_entry;
 mod create_rating;
+mod reset_password;
 mod update_entry;
 
 pub mod prelude {
     pub use super::{
         archive_comments::*, archive_entries::*, archive_events::*, archive_ratings::*,
-        change_user_role::*, create_entry::*, create_rating::*, update_entry::*,
+        change_user_role::*, create_entry::*, create_rating::*, reset_password::*, update_entry::*,
     };
 }
 
@@ -179,6 +180,11 @@ mod tests {
                     ..Default::default()
                 };
                 self.query_entries(&query)
+            }
+
+            pub fn create_user_from_email(self: &EnvFixture, email: &str) -> String {
+                usecases::create_user_from_email(&*self.db_connections.exclusive().unwrap(), email)
+                    .unwrap()
             }
         }